backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
bcrypt
passlib
python-multipart
itsdangerous
httpx
python-jose
